refact: 프로필 뷰 하나로 합침
